pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metisos-arc-core"
version = "0.1.0"
authors = [
    {name = "Metis AI Research", email = "cjohnson@metisos.com"},
]
description = "Adaptive Recursive Consciousness Engine - Modular continual learning for language models"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research", 
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = [
    "artificial intelligence",
    "machine learning", 
    "continual learning",
    "language models",
    "cognitive science",
    "neural networks",
    "biologically inspired",
    "recursive consciousness",
    "adaptive learning",
    "memory systems",
]
dependencies = [
    "torch>=1.12.0",
    "transformers>=4.20.0",
    "peft>=0.3.0",
    "numpy>=1.21.0",
    "click>=8.0.0",
    "PyYAML>=6.0",
    "huggingface_hub>=0.14.0",
    "tqdm>=4.64.0",
    "safetensors>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
    "pre-commit>=2.15",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
    "sphinx-autodoc-typehints>=1.12",
]
gpu = [
    "torch[cuda]>=1.12.0",
]
apple = [
    "torch>=1.12.0",
]

[project.urls]
Homepage = "https://github.com/metisai/arc-core"
Documentation = "https://arc-core.readthedocs.io/"
Repository = "https://github.com/metisai/arc-core"
"Bug Tracker" = "https://github.com/metisai/arc-core/issues"

[project.scripts]
arc = "cli.arc:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["arc_core*", "cli*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
arc_core = ["packs/**/*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
